pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sian-process-intelligence"
version = "0.1.0"
description = "Process Intelligence para processos organizacionais (Contas a Pagar e outros)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "SIAN" }]
dependencies = [
    "pandas>=2.0",
    "packaging>=21.0",
    "pm4py>=2.7",
    "simpy>=4.1",
    "google-cloud-bigquery>=3.0",
    "db-dtypes>=1.0",
    "pyarrow>=14.0",
    "duckdb>=0.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest-cov",
    "pytest-xdist",
    "ipython",
    "jupyter",
]
notebooks = [
    "jupyter",
    "jupyterlab",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src/app"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "def __repr__", "raise NotImplementedError"]
